pyproject.toml
[tool.poetry]
name = "duck-iceberg-demo"
version = "0.1.0"
description = ""
authors = ["Steven Wang <steven@luabase.com>"]
readme = "README.md"
packages = [{include = "duck_iceberg_demo"}]

[tool.poetry.dependencies]
python = "^3.10"
duckdb = "^1.0.0"
pyiceberg = {extras = ["gcsfs"], version = "^0.6.1"}
sqlalchemy = "^2.0.31"
pyarrow = "^16.1.0"
pandas = "^2.2.2"
psycopg2 = "^2.9.9"
flask = "^3.0.3"
google-cloud-secret-manager = "^2.20.1"
boto3 = "^1.34.157"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
